agents/orchestrator/requirements.txt
fastapi
uvicorn
requests
langchain
langchain-google-genai
python-dotenv
